Solve the duplicate temperature report data problem
The DWIN touch display support is attached to the second serial port, which the version of Marlin their firmware was based on assumed would accept temperature reports. Creality "fixed" this by tweaking the SERIAL_OUT macro in serial.h to send the same text to the default serial port twice which stopped the temperature reports going to the display (which would presumably cause problems), but created the temperature reporting problem.
This change undoes the SERIAL_OUT macro hacking, and instead puts a guard around the standard temperature reporting to not send to both ports if we have the serial-attached DWIN touchscreen.